requirements.txt
Flask==2.3.3
intasend-python==1.1.2
